refactor: move test util package into isolated one
