+++ internal/generator/pedersen/template/pedersen.test.go.tmpl
@@ -2,7 +2,7 @@ import (
 	"fmt"
 	curve "github.com/consensys/gnark-crypto/ecc/{{.Name}}"
 	"github.com/consensys/gnark-crypto/ecc/{{.Name}}/fr"
-	"github.com/consensys/gnark-crypto/utils"
+	"github.com/consensys/gnark-crypto/utils/testutils"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -52,7 +52,7 @@ func testCommit(t *testing.T, values ...interface{}) {
 	var (
 		pk              []ProvingKey
 		vk              VerifyingKey
-		err 			error
+		err             error
 		commitment, pok curve.G1Affine
 	)
 	valuesFr := interfaceSliceToFrSlice(t, values...)
@@ -103,7 +103,7 @@ func TestFoldProofs(t *testing.T) {
 		assert.Equal(t, pok, pokFolded)
 	})
 
-	t.Run("run empty", func (t *testing.T) {
+	t.Run("run empty", func(t *testing.T) {
 		var foldedCommitment curve.G1Affine
 		pok, err := BatchProve([]ProvingKey{}, [][]fr.Element{}, []byte("test"))
 		assert.NoError(t, err)
@@ -160,8 +160,8 @@ func TestMarshal(t *testing.T) {
 	vk.GRootSigmaNeg, err = randomOnG2()
 	assert.NoError(t, err)
 
-	t.Run("ProvingKey -> Bytes -> ProvingKey must remain identical.", utils.SerializationRoundTrip(&pk))
-	t.Run("ProvingKey -> Bytes (raw) -> ProvingKey must remain identical.", utils.SerializationRoundTripRaw(&pk))
-	t.Run("VerifyingKey -> Bytes -> VerifyingKey must remain identical.", utils.SerializationRoundTrip(&vk))
-	t.Run("VerifyingKey -> Bytes (raw) -> ProvingKey must remain identical.", utils.SerializationRoundTripRaw(&vk))
+	t.Run("ProvingKey -> Bytes -> ProvingKey must remain identical.", testutils.SerializationRoundTrip(&pk))
+	t.Run("ProvingKey -> Bytes (raw) -> ProvingKey must remain identical.", testutils.SerializationRoundTripRaw(&pk))
+	t.Run("VerifyingKey -> Bytes -> VerifyingKey must remain identical.", testutils.SerializationRoundTrip(&vk))
+	t.Run("VerifyingKey -> Bytes (raw) -> ProvingKey must remain identical.", testutils.SerializationRoundTripRaw(&vk))
 }
