+++ utils/testutils/testing.go
@@ -1,4 +1,4 @@
-package utils
+package testutils
 
 import (
 	"bytes"
